requirements.txt
affine==2.4.0
aiodns==3.0.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Brotli==1.0.9
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
DateTime==5.0
frozenlist==1.3.3
geojson==3.0.0
idna==3.4
multidict==6.0.4
numpy==1.24.2
packaging==23.0
pandas==1.5.3
pycares==4.3.0
pycparser==2.21
pyparsing==3.0.9
pyproj==3.4.1
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2022.7.1
rasterio==1.3.6
requests==2.28.2
rioxarray==0.13.3
six==1.16.0
snuggs==1.4.7
urllib3==1.26.14
xarray==2023.2.0
yarl==1.8.2
zope.interface==5.5.2
